Fix test compilation after refactorings
