+++ comm/src/test/scala/coop/rchain/comm/connect/FindAndConnectSpec.scala
@@ -10,7 +10,8 @@ import org.scalatest.enablers.Containing
 import cats._, cats.data._, cats.implicits._
 import coop.rchain.catscontrib._, Catscontrib._, ski._
 import coop.rchain.shared._
-import coop.rchain.comm.transport._, CommMessages._
+import coop.rchain.comm.transport._
+import coop.rchain.comm.rp.ProtocolHelper._
 
 class FindAndConnectSpec extends FunSpec with Matchers with BeforeAndAfterEach with AppendedClues {
